% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check.TriBounds.R
\name{Check.TriBounds}
\alias{Check.TriBounds}
\title{Check the bounds for the third pair's correlation in the trivariate setting}
\usage{
Check.TriBounds(cor, pv, low_bdd, up_bdd, ord = c(1, 2, 3), output = T)
}
\arguments{
\item{cor}{A vector that contains the sorted correlations}

\item{pv}{A vector contains the sorted proportions}

\item{low_bdd}{A vector contains the sorted lower bounds}

\item{up_bdd}{A vector contains the sorted upper bounds}

\item{ord}{A vector contains the order for each variable}

\item{output}{An option for whether gives an output}
}
\value{
An error message will be displayed when the third correlation is not within the trivariate bounds.
}
\description{
This function checks the extra bounds in the trivariate extension for this method.
}
\examples{
 f1 = function(n){rnorm(n)}
 cor_mat = matrix(c(1,.49,.1, .49, 1, -.4, .1, -.4, 1), nrow = 3)
 l = Find.Order(list(f1, f1, f1), cor_mat)
 cor = l[[7]]
 pv = l[[6]]
 low_bdd = l[[3]]
 up_bdd = l[[4]]
 ord = l[[1]]
 Check.TriBounds(cor, pv, low_bdd, up_bdd, ord, TRUE)

}
