% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute.SortProp.R
\name{Compute.SortProp}
\alias{Compute.SortProp}
\title{Compute the proportion needed to be sorted between two variables}
\usage{
Compute.SortProp(cor_mat, pairbounds)
}
\arguments{
\item{cor_mat}{Specified correlation matrix}

\item{pairbounds}{Calculated correlation bounds from the function \code{Compute.PairBounds}
based on Demirtas and Hedeker (2011).}
}
\value{
It will return the proportions needed to be sorted.
}
\description{
This function calculates the proportion needed to be sorted for the targeted correlation.
}
\examples{
 # Calculate correlation bounds of each pairs using the function Compute.PairBounds.
 # Three variables of normal,rnorm(n), Poisson,rpois(n,1), Binomial,
 # rbinom(n,10,0.5) are generated.
 # Define the 3x3 correlation matrix, cor_mat having the correlation coefficient
 # 0.5 for all three pairs.

 f1 = function(n){rnorm(n)}
 f2 = function(n){rpois(n,1)}
 f3 = function(n){rbinom(n,10,0.5)}
 pairbound = Compute.PairBounds(list(f1,f2,f3))
 cor_mat = matrix(0.5, 3, 3) + diag(0.5,3)
 Compute.SortProp(cor_mat, pairbound)



}
\seealso{
\code{\link{Compute.PairBounds}}
}
