% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorOrder.ToVar.R
\name{CorOrder.ToVar}
\alias{CorOrder.ToVar}
\title{Correlations order to the sorting order of variables}
\usage{
CorOrder.ToVar(cor_name)
}
\arguments{
\item{cor_name}{A list of correlation names following a certain order}
}
\value{
It will return an order for variables.
}
\description{
This function returns an order of variables according the order of correlations.
}
