% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find.Order.R
\name{Find.Order}
\alias{Find.Order}
\title{Find the order for sorting}
\usage{
Find.Order(lst, cor_mat)
}
\arguments{
\item{lst}{A list of functions which generate data under specified marginal distributions separately}

\item{cor_mat}{Specified correlation matrix}
}
\value{
A list is returned which contains the sorting order, ordered variables,  ordered lower bounds, ordered upper bounds, names of ordered correlation, ordered proportion vectors, ordered proportion vectors and ordered correlations.
  \item{\code{order}}{The order for sorting which achieves the maximum bounds}
  \item{\code{ordered list}}{An ordered list which generates the random variables}
  \item{\code{ordered lower bounds}}{The lower bounds corresponding to the ordered variables}
  \item{\code{ordered upper bounds}}{The upper bounds corresponding to the ordered variables}
  \item{\code{names of ordered correlation}}{Names of ordered correlation}
  \item{\code{ordered proportion vectors}}{The sorting proportion corresponding to the ordered variables}
  \item{\code{ordered correlations}}{Ordered correlations}
}
\description{
This function finds the order for trivariate sorting
}
\examples{
 f1 = function(n){rnorm(n)}
 cor_mat = matrix(c(1,.49,.1, .49, 1, -.4, .1, -.4, 1), nrow = 3)
 Find.Order(list(f1, f1, f1), cor_mat)



}
\seealso{
\code{\link{Compute.SortProp}}, \code{\link{Compute.PairBounds}}, \code{\link{CorOrder.ToVar}}
}
