% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.Correlation.R
\name{Validate.Correlation}
\alias{Validate.Correlation}
\title{Validate the correlation matrix}
\usage{
Validate.Correlation(cor_mat, pairbounds)
}
\arguments{
\item{cor_mat}{Specified correlation matrix}

\item{pairbounds}{Computed correlation bounds from the package \code{Compute.PairBounds}.}
}
\value{
An error message will be displayed when the matrix doesn't satisfy any of the three
 conditions.
}
\description{
This function validates the correlation matrix whether meets the conditions for this method (1) positive semi-definite;
 (2) correlation coefficient is not 0; (3) is within correlation bounds
 calculated from the package \code{Compute.PairBounds} by Demirtas and Hedeker(2011).
}
\examples{

 f1 = function(n){rnorm(n)}
 f2 = function(n){rpois(n,1)}
 f3 = function(n){rbinom(n,10,0.5)}
 pairbound = Compute.PairBounds(list(f1,f2,f3))
 cor_mat = matrix(0.5, 3, 3) + diag(0.5,3)
 Compute.SortProp(cor_mat, pairbound)
 Validate.Correlation(cor_mat = cor_mat,pairbound)

}
\seealso{
\code{\link{Compute.PairBounds}}
}
